refactor: move models & validators to seperate files
