+++ internal/provider/publiccloud/instances_datasource.go
@@ -9,8 +9,6 @@ import (
 	"github.com/hashicorp/terraform-plugin-framework/datasource"
 	"github.com/hashicorp/terraform-plugin-framework/datasource/schema"
 	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
-	"github.com/hashicorp/terraform-plugin-framework/types"
-	"github.com/hashicorp/terraform-plugin-framework/types/basetypes"
 	"github.com/hashicorp/terraform-plugin-log/tflog"
 	"github.com/leaseweb/leaseweb-go-sdk/publicCloud"
 	"github.com/leaseweb/terraform-provider-leaseweb/internal/provider/client"
@@ -21,94 +19,6 @@ var (
 	_ datasource.DataSource              = &InstancesDataSource{}
 	_ datasource.DataSourceWithConfigure = &InstancesDataSource{}
 )
-
-type dataSourceModelContract struct {
-	BillingFrequency types.Int64  `tfsdk:"billing_frequency"`
-	Term             types.Int64  `tfsdk:"term"`
-	Type             types.String `tfsdk:"type"`
-	EndsAt           types.String `tfsdk:"ends_at"`
-	State            types.String `tfsdk:"state"`
-}
-
-func newDataSourceModelContract(sdkContract publicCloud.Contract) dataSourceModelContract {
-	return dataSourceModelContract{
-		BillingFrequency: basetypes.NewInt64Value(int64(sdkContract.GetBillingFrequency())),
-		Term:             basetypes.NewInt64Value(int64(sdkContract.GetTerm())),
-		Type:             basetypes.NewStringValue(string(sdkContract.GetType())),
-		EndsAt:           utils.AdaptNullableTimeToStringValue(sdkContract.EndsAt.Get()),
-		State:            basetypes.NewStringValue(string(sdkContract.GetState())),
-	}
-}
-
-type dataSourceModelInstance struct {
-	ID                  types.String            `tfsdk:"id"`
-	Region              types.String            `tfsdk:"region"`
-	Reference           types.String            `tfsdk:"reference"`
-	Image               dataSourceModelImage    `tfsdk:"image"`
-	State               types.String            `tfsdk:"state"`
-	Type                types.String            `tfsdk:"type"`
-	RootDiskSize        types.Int64             `tfsdk:"root_disk_size"`
-	RootDiskStorageType types.String            `tfsdk:"root_disk_storage_type"`
-	Ips                 []dataSourceModelIp     `tfsdk:"ips"`
-	Contract            dataSourceModelContract `tfsdk:"contract"`
-	MarketAppId         types.String            `tfsdk:"market_app_id"`
-}
-
-func adaptSdkInstanceToDatasourceInstance(sdkInstance publicCloud.Instance) dataSourceModelInstance {
-	var ips []dataSourceModelIp
-	for _, ip := range sdkInstance.Ips {
-		ips = append(ips, adaptSdkIpToDatasourceIp(ip))
-	}
-
-	return dataSourceModelInstance{
-		ID:                  basetypes.NewStringValue(sdkInstance.GetId()),
-		Region:              basetypes.NewStringValue(string(sdkInstance.GetRegion())),
-		Reference:           basetypes.NewStringPointerValue(sdkInstance.Reference.Get()),
-		Image:               adaptSdkImageToDatasourceImage(sdkInstance.GetImage()),
-		State:               basetypes.NewStringValue(string(sdkInstance.GetState())),
-		Type:                basetypes.NewStringValue(string(sdkInstance.GetType())),
-		RootDiskSize:        basetypes.NewInt64Value(int64(sdkInstance.GetRootDiskSize())),
-		RootDiskStorageType: basetypes.NewStringValue(string(sdkInstance.GetRootDiskStorageType())),
-		Ips:                 ips,
-		Contract:            newDataSourceModelContract(sdkInstance.GetContract()),
-		MarketAppId:         basetypes.NewStringPointerValue(sdkInstance.MarketAppId.Get()),
-	}
-}
-
-type dataSourceModelImage struct {
-	Id types.String `tfsdk:"id"`
-}
-
-func adaptSdkImageToDatasourceImage(sdkImage publicCloud.Image) dataSourceModelImage {
-	return dataSourceModelImage{
-		Id: basetypes.NewStringValue(sdkImage.Id),
-	}
-}
-
-type dataSourceModelIp struct {
-	Ip types.String `tfsdk:"ip"`
-}
-
-func adaptSdkIpToDatasourceIp(sdkIp publicCloud.Ip) dataSourceModelIp {
-	return dataSourceModelIp{
-		Ip: basetypes.NewStringValue(sdkIp.Ip),
-	}
-}
-
-type dataSourceModelInstances struct {
-	Instances []dataSourceModelInstance `tfsdk:"instances"`
-}
-
-func adaptSdkInstancesToDatasourceInstances(sdkInstances []publicCloud.Instance) dataSourceModelInstances {
-	var instances dataSourceModelInstances
-
-	for _, sdkInstance := range sdkInstances {
-		instance := adaptSdkInstanceToDatasourceInstance(sdkInstance)
-		instances.Instances = append(instances.Instances, instance)
-	}
-
-	return instances
-}
 
 func getAllInstances(ctx context.Context, api publicCloud.PublicCloudAPI) (
 	[]publicCloud.Instance,
